requirements.txt
discord.py==2.3.2
PyNaCl==1.5.0
Flask==3.0.0
python-dotenv==1.0.0